Changed data representation and supply peak value
1. The light, humidity and temperature raw data have been converted
correctly in luminance, relative humidity (percentage) and Celsius
degree respectively.
They are printed out. Strangely the light measure output is always 0. To
be fixed.

2. Fixed the peak power consumption value in the pollution sensor: it
has been to multiply by 5 and not by 9.

+++ sendPacket2/sky/general-sensing-data-structure.h
@@ -2,9 +2,9 @@
 #define GENERAL_SENSE_DATA_STRUCT_H 4212
 
 struct sensed_data {
- uint16_t light;
- uint16_t hum;
- uint16_t temp;
+ unsigned long light;
+ unsigned long hum;
+ unsigned long temp;
 };
 
 #endif
